requirements.txt
google-api-python-client>=2.121.0
google-auth>=2.33.0
google-auth-oauthlib>=1.2.1
python-telegram-bot==21.6
google-generativeai>=0.8.0
python-dotenv>=1.0.1
beautifulsoup4>=4.12.3
tenacity>=8.2.3
